added helper method for updating schema metadata
